code refactor and live mode
